requirements.txt
selenium>=4.18.0
webdriver-manager>=4.0.1
pydantic>=2.6.0
pyyaml>=6.0.1
Pillow>=10.0.0
click>=8.1.7
